requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dateutil

# Gemini API SDK
google-generativeai

# Translation
deep-translator
streamlit 
requests
